pyproject.toml
[tool.poetry]
name = "auxia"
version = "0.0.1"
description = "Backend do programa Auxia de treinamento de IA utilizando RLHF"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.11"}
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pydantic-settings = "^2.8.1"
motor = "^3.7.0"
pre-commit = "^4.1.0"
httpx = "^0.28.1"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
google-genai = "^1.5.0"


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
  "--strict-config",
  "--strict-markers",
  "--ignore=docs_src",
]
xfail_strict = true
junit_family = "xunit2"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
